removed withReduxStore since we now use a provider directly (#681)
