--- unlock-app/src/utils/withReduxStore.js
@@ -1,55 +1,0 @@
-import React from 'react'
-import configure from '../config'
-import createUnlockStore from '../createUnlockStore'
-
-const config = configure(global)
-
-const __NEXT_REDUX_STORE__ = '__NEXT_REDUX_STORE__'
-
-function getOrCreateStore(initialState) {
-  // Always make a new store if server, otherwise state is shared between requests
-  if (config.isServer) {
-    return createUnlockStore(initialState)
-  }
-
-  // Create store if unavailable on the client and set it on the window object
-  if (!window[__NEXT_REDUX_STORE__]) {
-    window[__NEXT_REDUX_STORE__] = createUnlockStore(initialState)
-  }
-  return window[__NEXT_REDUX_STORE__]
-}
-
-export default App => {
-  return class AppWithRedux extends React.Component {
-    static async getInitialProps(appContext) {
-      // Get or Create the store with `undefined` as initialState
-      // This allows you to set a custom default initialState
-      const reduxStore = getOrCreateStore()
-      // Provide the store to getInitialProps of pages
-      appContext.ctx.reduxStore = reduxStore
-
-      let appProps = {}
-      if (typeof App.getInitialProps === 'function') {
-        appProps = await App.getInitialProps(appContext)
-      }
-
-      return {
-        ...appProps,
-        initialReduxState: reduxStore.getState(),
-      }
-    }
-
-    /**
-     * initialReduxState is actually coming from the server side rendered content
-     * We want to "merge" it with the data that might be stored in localStorage
-     */
-    constructor(props, context) {
-      super(props, context)
-      this.reduxStore = getOrCreateStore(props.initialReduxState)
-    }
-
-    render() {
-      return <App {...this.props} reduxStore={this.reduxStore} />
-    }
-  }
-}
